+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Mon, 09 May 2022 16:30:15 GMT and should not be manually modified.
+This log was last generated on Tue, 10 May 2022 19:19:34 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 5.0.18
+
+Tue, 10 May 2022 19:19:34 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v8.6.0
 
 ## 5.0.17